systemd: drop 0006-kmod-static-nodes.service-Don-t-require-lib-modules.patch
In general, we want to check for kernel modules to be present before
invoking kmod to generate a tmpfiles file. We could maybe hardcode it to
/run/{booted,current}-system/kernel-modules/lib/modules/%v/modules.dep,
but we shouldn't do it by patching systemd. Instead, we can also
override the vanilla upstream unit from the NixOS module system.

Instead, look at
${config.system.modulesTree}/lib/modules/%v/modules.dep